refactor: make notebooks relative to source directory
[breaking change]: this commit reorganizes notebooks to be associated
with a source code repository.

this cleans up a lot of noise in the common case, where notebooks and
bookmarks are being created for a particular source code repository.

after this commit creating, listing, and opening bookmarks and notebooks
are tied to the source code repository root currently opened by neovim.

Signed-off-by: ldelossa <louis.delos@gmail.com>